# Improve performance of qt.Qobj by static version check