--- a/doc/changes/2557.misc
+++ b/doc/changes/2557.misc
@@ -1,0 +1,1 @@
+Improve performance of qt.Qobj by using static numpy version check