**Description**

We improve performance of `Qobj` construction by moving the numpy version check out of the `__init__`. On a microbenchmark
```
import numpy as np
import qutip as qt
x = np.array([1., 0.])
%timeit qt.Qobj(x)
```
performance improves from
```
23.5 μs ± 2.18 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```
to
```
8.24 μs ± 732 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```